requirements.txt
asgiref==3.10.0
distlib==0.3.6
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.12.2
platformdirs==3.8.0
psycopg2-binary==2.9.11
pyodbc==4.0.39
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.23.1
